Проведён рефакторинг
