+++ handlers/task_add_and_delete.go
@@ -3,26 +3,30 @@ package handlers
 import (
 	"encoding/json"
 	"go_final_project/database"
-	"go_final_project/model"
+	"go_final_project/task"
 	"log"
 	"net/http"
-
-	"github.com/jmoiron/sqlx"
 )
 
 // AddTaskHandler обрабатывает запросы на добавление новой задачи
-func AddTaskHandler(db *sqlx.DB) http.HandlerFunc {
+func AddTaskHandler(storage *database.Storage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 
-		var newTask model.Task
+		var newTask task.Task
 		if err := json.NewDecoder(r.Body).Decode(&newTask); err != nil {
 			log.Printf("Ошибка при декодировании JSON: %v", err)
 			http.Error(w, `{"error": "Ошибка при декодировании JSON"}`, http.StatusBadRequest)
 			return
 		}
 
-		taskID, err := database.InsertTask(db, newTask)
+		// Валидация данных
+		if newTask.Title == "" {
+			http.Error(w, `{"error": "Title cannot be empty"}`, http.StatusBadRequest)
+			return
+		}
+
+		taskID, err := storage.InsertTask(newTask)
 		if err != nil {
 			log.Printf("Ошибка при добавлении задачи в базу данных: %v", err)
 			http.Error(w, `{"error": "`+err.Error()+`"}`, http.StatusInternalServerError)
@@ -34,7 +38,7 @@ func AddTaskHandler(db *sqlx.DB) http.HandlerFunc {
 }
 
 // DeleteTaskHandler обработчик для удаления задачи по идентификатору
-func DeleteTaskHandler(db *sqlx.DB) http.HandlerFunc {
+func DeleteTaskHandler(storage *database.Storage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		taskID := r.URL.Query().Get("id")
 		if taskID == "" {
@@ -42,17 +46,15 @@ func DeleteTaskHandler(db *sqlx.DB) http.HandlerFunc {
 			return
 		}
 
-		err := database.RemoveTask(db, taskID)
+		err := storage.RemoveTask(taskID)
 		if err != nil {
-			if err.Error() == "task not found" {
+			if err.Error() == "задача не найдена" {
 				http.Error(w, `{"error": "Задача не найдена"}`, http.StatusNotFound)
 			} else {
 				http.Error(w, `{"error": "Внутренняя ошибка сервера"}`, http.StatusInternalServerError)
 			}
 			return
 		}
-
-		//json.NewEncoder(w).Encode(map[string]interface{}{"status": "success"})
 
 		// Возвращаем пустой ответ, так как тесты ожидают пустой результат
 		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
