+++ handlers/task_get.go
@@ -5,14 +5,12 @@ import (
 	"go_final_project/database"
 	"log"
 	"net/http"
-
-	"github.com/jmoiron/sqlx"
 )
 
 const TaskLimit = 50
 
 // GetTaskHandler обработчик для получения задачи по идентификатору
-func GetTaskHandler(db *sqlx.DB) http.HandlerFunc {
+func GetTaskHandler(storage *database.Storage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		taskID := r.URL.Query().Get("id")
 		if taskID == "" {
@@ -20,9 +18,9 @@ func GetTaskHandler(db *sqlx.DB) http.HandlerFunc {
 			return
 		}
 
-		task, err := database.FetchTaskByID(db, taskID)
+		task, err := storage.GetTask(taskID)
 		if err != nil {
-			if err.Error() == "task not found" {
+			if err.Error() == "задача не найдена" {
 				http.Error(w, `{"error": "Задача не найдена"}`, http.StatusNotFound)
 			} else {
 				log.Printf("Ошибка при получении задачи: %v", err)
@@ -36,12 +34,12 @@ func GetTaskHandler(db *sqlx.DB) http.HandlerFunc {
 }
 
 // GetTasksHandler обработчик для получения списка задач с возможностью поиска
-func GetTasksHandler(db *sqlx.DB) http.HandlerFunc {
+func GetTasksHandler(storage *database.Storage) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 
 		searchTerm := r.URL.Query().Get("search")
-		tasks, err := database.FetchTasks(db, searchTerm, TaskLimit)
+		tasks, err := storage.GetTasks(searchTerm, TaskLimit)
 		if err != nil {
 			log.Printf("Ошибка при получении задач: %v", err)
 			http.Error(w, `{"error": "Ошибка выполнения запроса"}`, http.StatusInternalServerError)
